requirements.txt
# SafetyMindPro Backend Dependencies - Production Ready
# Python 3.13 Compatible

# Core FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.17
pydantic[email]>=2.10.0

# Database
sqlalchemy>=2.0.23
alembic>=1.13.0

# Graph and Network Analysis
networkx>=3.2.1

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.0

# YAML Configuration
pyyaml>=6.0.1

# CORS & Security
# python-cors>=1.0.0

# Utilities
python-dateutil>=2.8.2
